+++ WHATSNEW.md
@@ -1,3 +1,5 @@
+ + added mixbit.com plugin
+
 2013.08.06, Version 0.4.2
 
  + added v.qq.com plugin